fix(objectschema)!: omit id from schema returned by without + only
the methods without and only return new schema objects with the same id
that was found on the original. In terms of schema reuse, this creates
two different schemas with the same id which is technically invalid.
Additionally, it was not possible to set the id property on an object
schema that had properties defined. This problem originates from the
setattribute function which will always set the attribute on the
schema properties if they are defined. In the case of the object schema,
this is almost always the case. This changes the id function on the
object schema to always generate a new schema with the id set on the
object schema rather than its properties

BREAKING CHANGE: ObjectSchema.id() will always set the id on the root object
BREAKING CHANGE: ObjectSchema.without() will omit id from the return schema
BREAKING CHANGE: ObjectSchema.only() will omit id from the return schema